third_party/blink/web_tests/platform/generic/http/tests/security/webgl-cross-origin-ImageBitmap-blocked-expected.txt

WebGL's tex(Sub)Image2D should throw a SecurityError exception when the ImageBitmap is not origin clean.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGBA, gl.RGBA, gl.UNSIGNED_BYTE, bitmap) did not throw exception.
PASS gl.getError() is gl.NO_ERROR
PASS gl.texSubImage2D(gl.TEXTURE_2D, 0, 0, 0, gl.RGBA, gl.UNSIGNED_BYTE, bitmap) did not throw exception.
PASS gl.getError() is gl.NO_ERROR
PASS gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGBA, gl.RGBA, gl.UNSIGNED_BYTE, bitmap) threw exception SecurityError: Failed to execute 'texImage2D' on 'WebGLRenderingContext': The ImageBitmap contains cross-origin data, and may not be loaded..
PASS gl.getError() is gl.NO_ERROR
PASS gl.texSubImage2D(gl.TEXTURE_2D, 0, 0, 0, gl.RGBA, gl.UNSIGNED_BYTE, bitmap) threw exception SecurityError: Failed to execute 'texSubImage2D' on 'WebGLRenderingContext': The ImageBitmap contains cross-origin data, and may not be loaded..
PASS gl.getError() is gl.NO_ERROR
PASS successfullyParsed is true

TEST COMPLETE
